heads spawn and can be destroyed

+++ Assets/ScriptableObjects/ELC_EnemySO.cs
@@ -9,7 +9,7 @@ public class ELC_EnemySO : ScriptableObject
     public string Name;
 
     [Header("Health")]
-    public float MaxHealth;
+    public int MaxHealth;
 
     [Header("Movements")]
     public float MovementSpeed;
